fixed compilation errors

+++ Backend/Project/Helpers/Token/ITokenProvider.cs
@@ -4,6 +4,6 @@ namespace WebApi.Helpers.Token
 {
     public interface ITokenProvider
     {
-        string MakeToken(IdentityUser<long> user, TokenConfiguration configuration);
+        string MakeToken(IdentityUser<int> user, TokenConfiguration configuration);
     }
 }
